requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
alembic==0.9.10           # via flask-migrate
argh==0.26.2              # via watchdog
backcall==0.1.0           # via ipython
blinker==1.4              # via flask-mail
click==6.7                # via flask, rq
decorator==4.3.0          # via ipython, traitlets
flask-login==0.4.1
flask-mail==0.9.1
flask-migrate==2.2.1
flask-script==2.0.6
flask-sqlalchemy==2.3.2
flask-wtf==0.14.2
flask==1.0.2
gunicorn==19.9.0
ipython-genutils==0.2.0   # via traitlets
ipython==6.4.0
itsdangerous==0.24        # via flask
jedi==0.12.1              # via ipython
jinja2==2.10              # via flask
mako==1.0.7               # via alembic
markupsafe==1.0           # via jinja2, mako
parso==0.3.1              # via jedi
pathtools==0.1.2          # via watchdog
pexpect==4.6.0            # via ipython
pickleshare==0.7.4        # via ipython
prompt-toolkit==1.0.15    # via ipython
psycopg2==2.7.5
ptyprocess==0.6.0         # via pexpect
pygments==2.2.0           # via ipython
pyjwt==1.6.4
python-dateutil==2.7.3    # via alembic
python-editor==1.0.3      # via alembic
pyyaml==3.13              # via watchdog
redis==2.10.6             # via rq
rq==0.12.0
simplegeneric==0.8.1      # via ipython
simplejson==3.16.0
six==1.11.0               # via prompt-toolkit, python-dateutil, traitlets
sqlalchemy==1.2.9         # via alembic, flask-sqlalchemy
traitlets==4.3.2          # via ipython
watchdog==0.8.3
wcwidth==0.1.7            # via prompt-toolkit
werkzeug==0.14.1          # via flask
wtforms==2.2.1            # via flask-wtf
